Testing Scenario-Aggregated SVD on 0D Reactor

+++ romnet/__init__.py
@@ -11,7 +11,7 @@ from . import utils
 # TensorFlow ------------------------------------------------------------------
 tf_setup = {
     'EPSILON':              1.e-15,
-    'DTYPE':                'float64',
+    'DTYPE':                'float32',
     'NUM_THREADS':          0,
     'TF_CPP_MIN_LOG_LEVEL': '3'
 }
